LayoutTests/http/tests/paymentrequest/updateWith-shippingOptions.https-expected.txt

CONSOLE MESSAGE: WebKit currently uses the first shipping option even if other shipping options are marked as selected.

PASS Calling `updateWith` with a new `shippingOptions` without `requestShipping` should not update any values.
FAIL Calling `updateWith` with a new `shippingOptions` should not update any other values. assert_equals: selected shipping option should change expected "shippingId4" but got "shippingId3"
